requirements.txt
# SmartLearn AI — Python Backend Dependencies
# Install with: pip install -r requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
bcrypt==4.1.3
PyJWT==2.8.0
httpx==0.27.0
python-dotenv==1.0.1
